refactor file structure and cmake
