+++ kernel/src/arch/x86/vga/vga.c
